fix: useWhiteboard
