+++ packages/roomkit-react/src/Prebuilt/components/Footer/WhiteboardToggle.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { selectLocalPeer, selectWhiteboard, useHMSActions, useHMSStore } from '@100mslive/react-sdk';
+import { useWhiteboard } from '@100mslive/react-sdk';
 import { PencilDrawIcon } from '@100mslive/react-icons';
 import { Tooltip } from '../../..';
 // @ts-ignore: No implicit Any
@@ -8,26 +8,25 @@ import IconButton from '../../IconButton';
 import { ToastManager } from '../Toast/ToastManager';
 
 export const WhiteboardToggle = () => {
-  const localPeerUserId = useHMSStore(selectLocalPeer)?.customerUserId;
-  const open = useHMSStore(selectWhiteboard)?.open;
-  const isOwner = useHMSStore(selectWhiteboard)?.owner === localPeerUserId;
-  const actions = useHMSActions().interactivityCenter.whiteboard;
-
-  const toggle = async () => {
-    try {
-      if (open) {
-        isOwner && (await actions.close());
-      } else {
-        await actions.open();
-      }
-    } catch (error) {
-      ToastManager.addToast({ title: (error as Error).message, variant: 'error' });
-    }
-  };
+  const { toggle, open, isOwner } = useWhiteboard();
+  if (!toggle) {
+    return null;
+  }
 
   return (
     <Tooltip key="whiteboard" title={`${open ? 'Close' : 'Open'} Whiteboard`}>
-      <IconButton onClick={toggle} active={!open} disabled={open && !isOwner} data-testid="whiteboard_btn">
+      <IconButton
+        onClick={async () => {
+          try {
+            await toggle();
+          } catch (error) {
+            ToastManager.addToast({ title: (error as Error).message, variant: 'error' });
+          }
+        }}
+        active={!open}
+        disabled={open && !isOwner}
+        data-testid="whiteboard_btn"
+      >
         <PencilDrawIcon />
       </IconButton>
     </Tooltip>
